// src/app/classes/app-state.interface.ts
export interface StateEntity<T> {
  order?: string[];
  error: LookupTable<any>;
  loading: LookupTable<boolean>;
  byId: LookupTable<T>;
}
// ...
export interface LookupTable<T> {
  [id: string]: T;
}